refactor(ui): split admin modal components
